+++ common/src/gametest/java/com/grim3212/assorted/decor/gametest/DecorGameTests.java
@@ -24,6 +24,7 @@ public final class DecorGameTests {
         EntityDecorationTests.register(out);
         MenuTests.register(out);
         BlockDecorationTests.register(out);
+        GateTests.register(out);
         TooltipTests.register(out);
         AssetTests.register(out);
     }