+++ common/src/main/java/com/grim3212/assorted/decor/config/DecorCommonConfig.java
@@ -19,6 +19,7 @@ public class DecorCommonConfig {
     public final Supplier<Boolean> cageEnabled;
     public final Supplier<Boolean> planterPotEnabled;
     public final Supplier<Boolean> extrasEnabled;
+    public final Supplier<Boolean> gatesEnabled;
 
     public final Supplier<Boolean> colorizerConsumeBlock;
     public final Supplier<Integer> colorizerBrushCount;
@@ -46,6 +47,7 @@ public class DecorCommonConfig {
         cageEnabled = builder.defineBoolean("parts.cageEnabled", true, "Set this to true if the Cage should be craftable and visible in the creative tab");
         planterPotEnabled = builder.defineBoolean("parts.planterPotEnabled", true, "Set this to true if the Planter pot should be craftable and visible in the creative tab");
         extrasEnabled = builder.defineBoolean("parts.extrasEnabled", true, "Set this to true if the extras (chain link fence and new doors) should be craftable and visible in the creative tab");
+        gatesEnabled = builder.defineBoolean("parts.gatesEnabled", true, "Set this to true if the castle gate and garage door, and the trumpet and remote that open them, should be craftable and visible in the creative tab");
 
         colorizerConsumeBlock = builder.defineBoolean("colorizer.colorizerConsumeBlock", true, "Set this to true if the colorizer brush should consume a block when using them");
         colorizerBrushCount = builder.defineInteger("colorizer.colorizerBrushCount", 16, 1, 400, "Set this to the amount of blocks that a brush will be able to colorize after grabbing a block");